+++ recipely-app/screens/AuthScreen.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import { StyleSheet, View, Text } from 'react-native';
+
+class AuthScreen extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <View>
+        <Text>Auth screen!</Text>
+      </View>
+    );
+  }
+}
+
+export default AuthScreen;
